requirements.txt
absl-py==1.3.0
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.3.8
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
boto3==1.28.40
botocore==1.31.40
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.5.24
cryptography==40.0.2
dataclasses-json==0.5.7
distro==1.8.0
duckdb==0.7.1
faiss-cpu==1.7.3
fastapi==0.95.1
filelock==3.8.2
fire==0.5.0
frozenlist==1.3.3
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.11.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langchain==0.0.166
llama-cpp-python==0.1.48
lz4==4.3.2
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==22.0
pandas==2.0.1
paper-qa==0.0.21
pdfminer.six==20221105
Pillow==9.5.0
posthog==3.0.1
prompt-toolkit==1.0.14
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.16.1
pygpt4all==1.1.0
pygptj==2.0.3
PyInquirer==1.0.3
pyllamacpp==2.1.3
pypdf==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2022.10.31
replicate==0.11.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-rust==1.5.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
threadpoolctl==3.1.0
tiktoken==0.3.2
tokenizers==0.13.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.2
yarl==1.8.2
zstandard==0.21.0
